bob-voice-updates/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets>=13.0
httpx==0.27.2
deepgram-sdk==3.7.7
elevenlabs==2.39.0
python-dotenv==1.0.1
chromadb==0.5.23
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
